requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
